refactor(HR): Auto Attendance
> moved all auto attendance settings from HR Settings to shift
> added shift in attendance and Employee Attendance Log
> reordered and cleaned up fields in HR Settings and Employee DocType
